Convert TipService to StoreKit 2

+++ DuffyFramework/TipOption.swift
@@ -11,7 +11,7 @@ import Foundation
 public struct TipOption {
     public let identifier: TipIdentifier
     public let formattedPrice: String
-    public let price: Double
+    public let price: Decimal
 }
 
 public enum TipIdentifier: String, CaseIterable {
